requirements.txt
# Python 3.11
numpy==1.26.4
torch==2.9.0
torchaudio==2.9.0
soxr==1.0.0
accelerate==1.10.1
transformers==4.53.3
nemo_toolkit[asr]==2.5.0
gradio==5.49.1